Backend/requirements.txt
asgiref==3.4.1
Django==3.1.14
djangorestframework==3.11.2
pytz==2021.3
sqlparse==0.4.2
django-cors-headers
